Extracted GameCard and added a test
